+++ pkg/utils/overriding/merging_test.go
@@ -7,238 +7,10 @@ import (
 	"testing"
 
 	dw "github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
-	attributesPkg "github.com/devfile/api/v2/pkg/attributes"
 	"github.com/stretchr/testify/assert"
-	"k8s.io/apimachinery/pkg/util/json"
 	yamlMachinery "k8s.io/apimachinery/pkg/util/yaml"
 	"sigs.k8s.io/yaml"
 )
-
-func TestBasicMerging(t *testing.T) {
-
-	tests := []struct {
-		name                    string
-		mainContent             *dw.DevWorkspaceTemplateSpecContent
-		parentFlattenedContent  *dw.DevWorkspaceTemplateSpecContent
-		pluginFlattenedContents []*dw.DevWorkspaceTemplateSpecContent
-		expected                *dw.DevWorkspaceTemplateSpecContent
-		wantErr                 *string
-	}{
-		{
-			name: "Basic Merging",
-			mainContent: &dw.DevWorkspaceTemplateSpecContent{
-				Variables: map[string]string{
-					"version1": "main",
-				},
-				Attributes: attributesPkg.Attributes{}.FromMap(map[string]interface{}{
-					"main": true,
-				}, nil),
-				Commands: []dw.Command{
-					{
-						Id: "mainCommand",
-						CommandUnion: dw.CommandUnion{
-							Exec: &dw.ExecCommand{
-								WorkingDir: "dir",
-							},
-						},
-					},
-				},
-				Components: []dw.Component{
-					{
-						Name: "mainComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Container: &dw.ContainerComponent{
-								Container: dw.Container{
-									Image: "image",
-								},
-							},
-						},
-					},
-					{
-						Name: "mainPluginComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Plugin: &dw.PluginComponent{
-								ImportReference: dw.ImportReference{
-									ImportReferenceUnion: dw.ImportReferenceUnion{
-										Uri: "uri",
-									},
-								},
-							},
-						},
-					},
-				},
-				Events: &dw.Events{
-					DevWorkspaceEvents: dw.DevWorkspaceEvents{
-						PostStop: []string{"post-stop-main"},
-					},
-				},
-			},
-			pluginFlattenedContents: []*dw.DevWorkspaceTemplateSpecContent{
-				{
-					Variables: map[string]string{
-						"version2": "plugin",
-					},
-					Attributes: attributesPkg.Attributes{}.FromMap(map[string]interface{}{
-						"plugin": true,
-					}, nil),
-					Commands: []dw.Command{
-						{
-							Id: "pluginCommand",
-							CommandUnion: dw.CommandUnion{
-								Exec: &dw.ExecCommand{
-									WorkingDir: "dir",
-								},
-							},
-						},
-					},
-					Components: []dw.Component{
-						{
-							Name: "pluginComponent",
-							ComponentUnion: dw.ComponentUnion{
-								Container: &dw.ContainerComponent{
-									Container: dw.Container{
-										Image: "image",
-									},
-								},
-							},
-						},
-					},
-					Events: &dw.Events{
-						DevWorkspaceEvents: dw.DevWorkspaceEvents{
-							PostStop: []string{"post-stop-plugin"},
-						},
-					},
-				},
-			},
-			parentFlattenedContent: &dw.DevWorkspaceTemplateSpecContent{
-				Variables: map[string]string{
-					"version3": "parent",
-				},
-				Attributes: attributesPkg.Attributes{}.FromMap(map[string]interface{}{
-					"parent": true,
-				}, nil),
-				Commands: []dw.Command{
-					{
-						Id: "parentCommand",
-						CommandUnion: dw.CommandUnion{
-							Exec: &dw.ExecCommand{
-								WorkingDir: "dir",
-							},
-						},
-					},
-				},
-				Components: []dw.Component{
-					{
-						Name: "parentComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Container: &dw.ContainerComponent{
-								Container: dw.Container{
-									Image: "image",
-								},
-							},
-						},
-					},
-				},
-				Events: &dw.Events{
-					DevWorkspaceEvents: dw.DevWorkspaceEvents{
-						PostStop:  []string{"post-stop-parent"},
-						PostStart: []string{"post-start-parent"},
-					},
-				},
-			},
-			expected: &dw.DevWorkspaceTemplateSpecContent{
-				Variables: map[string]string{
-					"version3": "parent",
-					"version2": "plugin",
-					"version1": "main",
-				},
-				Attributes: attributesPkg.Attributes{}.FromMap(map[string]interface{}{
-					"parent": true,
-					"plugin": true,
-					"main":   true,
-				}, nil),
-				Commands: []dw.Command{
-					{
-						Id: "parentCommand",
-						CommandUnion: dw.CommandUnion{
-							Exec: &dw.ExecCommand{
-								WorkingDir: "dir",
-							},
-						},
-					},
-					{
-						Id: "pluginCommand",
-						CommandUnion: dw.CommandUnion{
-							Exec: &dw.ExecCommand{
-								WorkingDir: "dir",
-							},
-						},
-					},
-					{
-						Id: "mainCommand",
-						CommandUnion: dw.CommandUnion{
-							Exec: &dw.ExecCommand{
-								WorkingDir: "dir",
-							},
-						},
-					},
-				},
-				Components: []dw.Component{
-					{
-						Name: "parentComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Container: &dw.ContainerComponent{
-								Container: dw.Container{
-									Image: "image",
-								},
-							},
-						},
-					},
-					{
-						Name: "pluginComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Container: &dw.ContainerComponent{
-								Container: dw.Container{
-									Image: "image",
-								},
-							},
-						},
-					},
-					{
-						Name: "mainComponent",
-						ComponentUnion: dw.ComponentUnion{
-							Container: &dw.ContainerComponent{
-								Container: dw.Container{
-									Image: "image",
-								},
-							},
-						},
-					},
-				},
-				Events: &dw.Events{
-					DevWorkspaceEvents: dw.DevWorkspaceEvents{
-						PreStart:  []string{},
-						PostStart: []string{"post-start-parent"},
-						PreStop:   []string{},
-						PostStop:  []string{"post-stop-main", "post-stop-parent", "post-stop-plugin"},
-					},
-				},
-			},
-			wantErr: nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mergedContent, err := MergeDevWorkspaceTemplateSpec(tt.mainContent, tt.parentFlattenedContent, tt.pluginFlattenedContents...)
-			if err != nil {
-				t.Error(err)
-				return
-			}
-
-			assert.Equal(t, tt.expected, mergedContent, "The two values should be the same.")
-		})
-	}
-}
 
 func mergingPatchTest(main, parent, expected []byte, expectedError string, plugins ...[]byte) func(t *testing.T) {
 	return func(t *testing.T) {
@@ -252,11 +24,7 @@ func mergingPatchTest(main, parent, expected []byte, expectedError string, plugi
 			return
 		}
 
-		resultJson, err := json.Marshal(result)
-		if err != nil {
-			t.Error(err)
-		}
-		resultYaml, err := yaml.JSONToYAML(resultJson)
+		resultYaml, err := yaml.Marshal(result)
 		if err != nil {
 			t.Error(err)
 		}
