+++ pkg/utils/overriding/test-fixtures/merges/no-duplicate/plugin.yaml
@@ -1,3 +1,7 @@
+variables:
+  version3: plugin
+attributes:
+  plugin: true
 components:
   - container:
       image: "aValue"
