+++ pkg/validation/variables/utils.go
@@ -8,17 +8,3 @@ func checkForInvalidError(invalidKeys map[string]bool, err error) {
 		}
 	}
 }
-
-// processInvalidKeys processes the invalid keys and return InvalidKeysError if present
-func processInvalidKeys(invalidKeys map[string]bool) error {
-	var invalidKeysArr []string
-	for key := range invalidKeys {
-		invalidKeysArr = append(invalidKeysArr, key)
-	}
-
-	if len(invalidKeysArr) > 0 {
-		return &InvalidKeysError{Keys: invalidKeysArr}
-	}
-
-	return nil
-}
